Диплом. Этап 1. Проверка 2.

+++ stats/stats-dto/src/main/java/ru/practicum/stats/dto/StatsHitDto.java
@@ -11,5 +11,5 @@ public class StatsHitDto {
 
     private String uri;
 
-    private Integer hits;
+    private long hits;
 }
